+++ playwright.config.ci.ts
@@ -10,17 +10,17 @@ const config: PlaywrightTestConfig = {
   },
   // Test directory
   testDir: "specs/e2e",
-  // If a test fails, retry it additional 2 times
-  retries: 5,
+  // If a test fails, retry it additional 5 times
+  retries: 3,
   // Artifacts folder where screenshots, videos, and traces are stored.
   outputDir: "test-results/",
-  workers: 5,
+  workers: 3,
   maxFailures: 10,
 
   use: {
     // Retry a test if its failing with enabled tracing. This allows you to analyse the DOM, console logs, network traffic etc.
     // More information: https://playwright.dev/docs/trace-viewer
-    trace: "retry-with-trace",
+    trace: "retain-on-failure",
     headless: true,
     viewport: { width: 1280, height: 720 },
     ignoreHTTPSErrors: true,