+++ specs/e2e/payments-adyen.spec.ts
@@ -41,6 +41,34 @@ test.describe("guest with Adyen", () => {
     await checkoutPage.page.reload()
     await checkoutPage.checkPaymentRecap("Credit card ending in ****")
   })
+
+  test("Checkout order using Credit Card with 3D Secure 2 authentication", async ({
+    checkoutPage,
+  }) => {
+    await checkoutPage.checkOrderSummary("Order Summary")
+
+    await checkoutPage.checkStep("Shipping", "open")
+
+    await checkoutPage.selectShippingMethod({ text: "Standard Shipping" })
+
+    await checkoutPage.save("Shipping")
+
+    await checkoutPage.selectPayment("adyen")
+
+    await checkoutPage.setPayment("adyen3DS")
+
+    await checkoutPage.save("Payment", undefined, true)
+
+    await checkoutPage.enter3DSecure({ type: "adyen", text: "password" })
+
+    await checkoutPage.page
+      .locator(`text=Thank you for your order!`)
+      .waitFor({ state: "visible", timeout: 10000 })
+
+    await checkoutPage.checkPaymentRecap(" ending in ****")
+    await checkoutPage.page.reload()
+    await checkoutPage.checkPaymentRecap(" ending in ****")
+  })
 })
 
 test.describe("guest with Adyen drop-in", () => {
@@ -84,27 +112,6 @@ test.describe("guest with Adyen drop-in", () => {
     await checkoutPage.checkPaymentRecap("PayPal ending in ****")
   })
 
-  test.skip("Checkout order with Klarna", async ({ checkoutPage }) => {
-    await checkoutPage.checkOrderSummary("Order Summary")
-
-    await checkoutPage.checkStep("Shipping", "open")
-
-    await checkoutPage.selectShippingMethod({ text: "Standard Shipping" })
-
-    await checkoutPage.save("Shipping")
-
-    await checkoutPage.selectPayment("adyen")
-
-    await checkoutPage.completePayment({
-      type: "adyen-dropin",
-      gateway: "klarna",
-    })
-
-    await checkoutPage.checkPaymentRecap("Klarna ending in ****")
-    await checkoutPage.page.reload()
-    await checkoutPage.checkPaymentRecap("Klarna ending in ****")
-  })
-
   test("Checkout order with Credit Card", async ({ checkoutPage }) => {
     await checkoutPage.checkOrderSummary("Order Summary")
 
@@ -124,6 +131,36 @@ test.describe("guest with Adyen drop-in", () => {
     await checkoutPage.checkPaymentRecap("Credit card ending in ****")
     await checkoutPage.page.reload()
     await checkoutPage.checkPaymentRecap("Credit card ending in ****")
+  })
+
+  test("Checkout order with Credit Card and 3D Secure", async ({
+    checkoutPage,
+  }) => {
+    await checkoutPage.checkOrderSummary("Order Summary")
+
+    await checkoutPage.checkStep("Shipping", "open")
+
+    await checkoutPage.selectShippingMethod({ text: "Standard Shipping" })
+
+    await checkoutPage.save("Shipping")
+
+    await checkoutPage.selectPayment("adyen")
+
+    await checkoutPage.completePayment({
+      type: "adyen-dropin",
+      gateway: "card3DS",
+    })
+
+    await checkoutPage.save("Payment", undefined, true)
+
+    await checkoutPage.enter3DSecure({ type: "adyen", text: "password" })
+
+    await checkoutPage.page
+      .locator(`text=Thank you for your order!`)
+      .waitFor({ state: "visible", timeout: 10000 })
+    await checkoutPage.checkPaymentRecap(" ending in ****")
+    await checkoutPage.page.reload()
+    await checkoutPage.checkPaymentRecap(" ending in ****")
   })
 })